+++ CHANGELOG.md
@@ -4,3 +4,9 @@
 ## 1.1.0
 - SDK resolve
 
+## 1.2.0
+- Import resolve
+
+## 1.3.0
+- Import resolve
+